Fixed bug with incorrect swiss group built for not-qualified players.
Example: let we have 22 players and we play Qualification+Playoff.

Main draw is 16, so 6 + 6 players should pass qualification and 10 are seeded.

After qualification there are 6 players who did not pass it.

In 13th version swissgroup hase been created with 6 players, w/o BYEs.
And it was an error.

Now, list of players are filled with BYEs until number of players become power
of 2.